+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 Change Log
 ==========
+
+Version 1.0.2 *(2015-02-23)*
+--------------------------
+ * switch to using a builder to configure the WordTokenizer
+
 Version 1.0.1 *(2015-02-21)*
 --------------------------
  *  disable spelling suggestions during mention suggestions
- *  Change target SDK to latest version (21) for spyglass.
- 
+ *  change target SDK to latest version (21) for spyglass.
+
 Version 1.0.0 *(2015-02-11)*
 --------------------------
  *  first public release